+++ composeApp/src/commonMain/kotlin/de/visualdigits/shipermansfriend/data/model/aisstreamio/data/BinaryAcknowledge.kt
@@ -7,9 +7,9 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class BinaryAcknowledge(
     @SerialName("Destinations") val destinations: Destinations? = Destinations(),
-    @SerialName("MessageID") val messageId: Int? = 0,
-    @SerialName("RepeatIndicator") val repeatIndicator: Int? = 0,
-    @SerialName("Spare") val spare: Int? = 0,
-    @SerialName("UserID") val mmsi: Int? = 0,
+    @SerialName("MessageID") val messageId: Long? = 0,
+    @SerialName("RepeatIndicator") val repeatIndicator: Long? = 0,
+    @SerialName("Spare") val spare: Long? = 0,
+    @SerialName("UserID") val mmsi: Long? = 0,
     @SerialName("Valid") val valid: Boolean? = false
 ) : AisMessageData
