+++ composeApp/src/commonMain/kotlin/de/visualdigits/shipermansfriend/data/model/aisstreamio/data/MultiSlotBinaryMessage.kt
@@ -8,15 +8,15 @@ import kotlinx.serialization.Serializable
 data class MultiSlotBinaryMessage(
     @SerialName("ApplicationID") val applicationID: ApplicationID? = null,
     @SerialName("ApplicationIDValid") val applicationIDValid: Boolean? = null,
-    @SerialName("CommunicationState") val communicationState: Int? = null,
+    @SerialName("CommunicationState") val communicationState: Long? = null,
     @SerialName("CommunicationStateIsItdma") val communicationStateIsItdma: Boolean? = null,
-    @SerialName("DestinationID") val destinationMmsi: Int? = null,
+    @SerialName("DestinationID") val destinationMmsi: Long? = null,
     @SerialName("DestinationIDValid") val destinationMmsiValid: Boolean? = null,
-    @SerialName("MessageID") val messageId: Int? = null,
+    @SerialName("MessageID") val messageId: Long? = null,
     @SerialName("Payload") val payload: String? = null,
-    @SerialName("RepeatIndicator") val repeatIndicator: Int? = null,
-    @SerialName("Spare1") val spare1: Int? = null,
-    @SerialName("Spare2") val spare2: Int? = null,
-    @SerialName("UserID") val mmsi: Int? = null,
+    @SerialName("RepeatIndicator") val repeatIndicator: Long? = null,
+    @SerialName("Spare1") val spare1: Long? = null,
+    @SerialName("Spare2") val spare2: Long? = null,
+    @SerialName("UserID") val mmsi: Long? = null,
     @SerialName("Valid") val valid: Boolean? = null
 ) : AisMessageData
