+++ composeApp/src/commonMain/kotlin/de/visualdigits/shipermansfriend/data/model/aisstreamio/data/SafetyAisMessageData.kt
@@ -5,8 +5,8 @@ import kotlinx.serialization.Serializable
 @Serializable
 sealed interface SafetyAisMessageData : AisMessageData {
 
-    val messageId :Int
-    val repeatIndicator :Int
+    val messageId :Long
+    val repeatIndicator :Long
     val mmsi: Long
     val valid :Boolean
     val text: String
