+++ composeApp/src/commonMain/kotlin/de/visualdigits/shipermansfriend/data/model/aisstreamio/data/SafetyBroadcastMessage.kt
@@ -6,9 +6,9 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class SafetyBroadcastMessage(
-    @SerialName("MessageID") override val messageId: Int,
-    @SerialName("RepeatIndicator") override val repeatIndicator: Int,
-    @SerialName("Spare") val spare: Int = 0,
+    @SerialName("MessageID") override val messageId: Long,
+    @SerialName("RepeatIndicator") override val repeatIndicator: Long,
+    @SerialName("Spare") val spare: Long = 0,
     @SerialName("Text") override val text: String,
     @SerialName("UserID") override val mmsi: Long,
     @SerialName("Valid") override val valid: Boolean
